// server/db/schema.ts
export const user = pgTable("user", {
  id: text("id").primaryKey(),
  name: text("name").notNull(),
  email: text("email").notNull().unique(),
  emailVerified: boolean("email_verified").default(false).notNull(),
  image: text("image"),
  createdAt: timestamp("created_at").defaultNow().notNull(),
  updatedAt: timestamp("updated_at")
    .defaultNow()
    .$onUpdate(() => new Date())
    .notNull(),
  companyName: text("company_name"),
  logoUrl: text("logo_url"),
  stripeCustomerId: text("stripe_customer_id"),
  subscriptionTier: text("subscription_tier")
    .default("free")
    .$type<"free" | "monthly" | "annual">(),
  subscriptionStatus: text("subscription_status").default("pending"),
  stripeSessionId: text("stripe_session_id"),
  estimatesThisMonth: integer("estimates_this_month").default(0).notNull(),
});